[Fix] #200 스크롤뷰 추가 및 레이아웃 수정
